restructure errors to be in their own file
