+++ game2048/__init__.py
@@ -21,4 +21,4 @@ socketio = SocketIO(app)
 login_manager.login_view = 'home'
 login_manager.login_message_category = 'info'
 
-from game2048 import routes, models
+from game2048 import routes, models, errors
