+++ game2048/errors.py
@@ -1,0 +1,11 @@
+from flask import render_template
+from game2048 import app, db
+
+@app.errorhandler(404)
+def not_found_error(error):
+    return render_template('404.html'), 404
+
+@app.errorhandler(500)
+def internal_error(error):
+    db.session.rollback()
+    return render_template('500.html'), 500
