+++ game2048/routes.py
@@ -58,15 +58,6 @@ def register():
     
     return render_template('register.html', title='Register', form=form)
 
-@app.errorhandler(404)
-def not_found_error(error):
-    return render_template('404.html'), 404
-
-@app.errorhandler(500)
-def internal_error(error):
-    db.session.rollback()
-    return render_template('500.html'), 500
-
 # ----------------------------------------------------------------
 # Anything below should have @login_required
 # ----------------------------------------------------------------
